Integrating Observer with the service and worker
