+++ backend/libs/exchanges/binance/observers/Cargo.toml
@@ -7,3 +7,25 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+errors = { path = "../../../errors" }
+types = { path = "../../../types" }
+config = { path = "../../../config" }
+rpc = { path = "../../../rpc" }
+
+entities = { path = "../../../rulesets/entities" }
+trade_observer = { path = "../../../rulesets/trade_observer" }
+symbol_recorder = { path = "../../../rulesets/symbol_recorder" }
+
+binance_clients = { path = "../clients" }
+binance_symbols = { path = "../symbols" }
+
+async-trait = "0.1.42"
+futures = "0.3.13"
+mongodb = {version = "1.1.1", default-features = false, features = ["async-std-runtime"]}
+nats = "0.8"
+rmp-serde = "0.15.4"
+serde_json = "1.0.62"
+slog = "2.7.0"
+tokio = "1.2.0"
+tokio-tungstenite = "0.13.0"
+serde = { version = "1.0.123", features = ["derive"] }
