+++ backend/libs/exchanges/binance/observers/src/constants.rs
@@ -1,0 +1,3 @@
+pub use ::binance_clients::constants::WS_ENDPOINT;
+pub use ::binance_symbols::constants::{SYMBOL_ADD_EVENT, SYMBOL_REMOVE_EVENT};
+pub const TRADE_OBSERVER_SUB_NAME: &str = "binance.trades";
