+++ backend/libs/exchanges/binance/observers/src/entities/book_ticker.rs
