--- backend/libs/exchanges/binance/streams/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-  println!("Hello, world!");
-}
