+++ backend/libs/exchanges/binance/symbols/src/lib.rs
@@ -1,4 +1,4 @@
-mod constants;
+pub mod constants;
 mod entities;
 pub mod fetcher;
 mod manager;
@@ -7,4 +7,4 @@ pub mod recorder;
 pub use ::symbol_fetcher::SymbolFetcher;
 pub use ::symbol_recorder::SymbolRecorder;
 
-pub use entities::ListSymbolStream;
+pub use self::entities::{ListSymbolStream, Symbol};
